restructure of functions in various logical files
